Fix ARRAY(*) assumed-size parameter binding (task #29)
User-reported (140-STATISTICS.hal's `CALL STATISTICS(DATA) ASSIGN(LO,
HI, MN);`, STATISTICS's own DATA parameter declared `ARRAY(*) SCALAR`):
"procedure/function call: MATRIX/VECTOR/ARRAY argument shape does not
match parameter 7".

Two compounding bugs:

1. symtab.c's EXT_ARRAY dimension-bound parsing never sign-extended
   from the field's real 16-bit width. An assumed-size parameter's own
   bound is encoded as a *negative* 16-bit sentinel (confirmed against
   two independent real-corpus cases with two different magnitudes:
   140-STATISTICS.hal's DATA is 0xFFF9 = -7, 141-VSUM.hal's V is
   0xFFFC = -4), previously misread as a huge unsigned "size" (65529)
   instead.

2. Even with that fixed, ensure_container() had no way to represent
   "this parameter's real size isn't known until a call actually
   supplies one". Fixed by having ensure_container() leave an
   assumed-size ARRAY parameter entirely unallocated, and
   bind_call_argument() allocate it directly from the caller's own
   actual argument shape instead of requiring a match against a
   declared size that doesn't exist for this kind of parameter.

Fixture: test_statistics.hal (140-STATISTICS.hal copied verbatim;
LO=MIN=10, HI=MAX=50, MEAN=SUM/SIZE=150/5=30, hand-derivable from
DATA's own INITIAL(10,20,30,40,50)).

+++ yaShuttle/yaHALMAT2/src/symtab.c
@@ -177,7 +177,28 @@ static void symtab_finalize(symtab_parse_state_t *st, halmat_symtab_t *out) {
                 if (dim_count > HALMAT_SYM_MAX_ARRAY_DIMS) dim_count = HALMAT_SYM_MAX_ARRAY_DIMS;
                 st->entries[i].array_dim_count = (int)dim_count;
                 for (uint32_t d = 0; d < dim_count && base + 1 + d < st->extarray_count; d++) {
-                    st->entries[i].array_dims[d] = (int)st->extarray[base + 1 + d];
+                    /* Each EXT_ARRAY dimension-bound entry is a 16-bit BIT
+                     * field (confirmed via the raw COMMON0.out dump --
+                     * "EXTuARRAY n BIT xxxx", always 4 hex digits), so it
+                     * must be sign-extended from int16_t, not read as a
+                     * plain unsigned value -- an `ARRAY(*)` (assumed-size)
+                     * parameter's own bound is encoded as a *negative*
+                     * 16-bit value (user-reported, 140-STATISTICS.hal's
+                     * `DECLARE DATA ARRAY(*) SCALAR;`: EXT_ARRAY entry
+                     * 0xFFF9 = -7, confirmed against a second real-corpus
+                     * case, 141-VSUM.hal's `ARRAY(*) VECTOR`: 0xFFFC = -4
+                     * -- a different magnitude, so the sentinel is "any
+                     * negative value," not one fixed constant). Read as
+                     * plain unsigned (the prior behavior), -7 became 65529
+                     * and -4 became 65532, both silently accepted as
+                     * enormous-but-"valid" array sizes by ensure_container
+                     * (falling through to its own generic 64-element
+                     * placeholder capacity instead), rather than being
+                     * recognized as the assumed-size marker they are --
+                     * see ensure_container()'s and bind_call_argument()'s
+                     * own comments (interp.c) for how a negative bound is
+                     * actually handled now. */
+                    st->entries[i].array_dims[d] = (int16_t)(uint16_t)st->extarray[base + 1 + d];
                 }
             }
             /* ARRAY-of-VECTOR/MATRIX (`ARRAY(n) VECTOR`, `ARRAY(n)
